方式
答案是

+++ src/star/mvc/common/sale_price.java
@@ -2,13 +2,10 @@ package star.mvc.common;
 
 public class sale_price {
 	public static String subsale(String price,String sale){
-		double price1,sale1,nowprice;
-		
-		price1 = Double.parseDouble(price);
-		sale1  = Double.parseDouble(sale);
-		
-		nowprice = price1*(sale1/10);
-		
+		float price1,sale1,nowprice;
+		price1 = Float.parseFloat(price);
+		sale1  = Float.parseFloat(sale);
+		nowprice = price1*(sale1/10f);
 		return String.valueOf(nowprice);
 	}
 }
